Move register methods to simplify generics
